refactor(test): rmove from definition to test file
